requirements.txt
psycopg2-binary~=2.9.9
Telethon~=1.37.0
openai~=1.45.0
aiogram~=3.13.1